+++ vNES/src/vnes/Mapper010.java
@@ -32,7 +32,7 @@ public class Mapper010 extends MapperDefault {
 
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
 
@@ -217,7 +217,7 @@ public class Mapper010 extends MapperDefault {
         super.mapperInternalStateSave(buf);
 
         // Version:
-        buf.putByte((short) 1);
+        buf.putByte((int) 1);
 
         // State:
         buf.putByte((byte) latchLo);
