+++ vNES/src/vnes/Mapper072.java
@@ -22,7 +22,7 @@ public class Mapper072 extends MapperDefault {
         super.init(nes);
     }
 
-    public void write(int address, short value) {
+    public void write(int address, int value) {
 
         if (address < 0x8000) {
             super.write(address, value);
